Revert "Albrja/mic 6186/intervention component (#527)"
This reverts commit 06ec6d38d16c7be4b96a6656fabfa22f84cf0a03.
