--- docs/source/api_reference/exposure/exposure.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_public_health.exposure.exposure
