--- docs/source/api_reference/exposure/index.rst
@@ -1,11 +1,0 @@
-=======================
-Exposure Modeling Tools
-=======================
-
-.. automodule:: vivarium_public_health.exposure
-
-.. toctree::
-   :maxdepth: 2
-   :glob:
-
-   *
