+++ docs/source/api_reference/risks/distributions.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium_public_health.risks.distributions
