--- docs/source/api_reference/treatment/intervention.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_public_health.treatment.intervention
