+++ src/vivarium_public_health/__init__.py
@@ -46,13 +46,7 @@ from vivarium_public_health.risks import (
     Risk,
     RiskEffect,
 )
-from vivarium_public_health.treatment import (
-    AbsoluteShift,
-    Intervention,
-    InterventionEffect,
-    LinearScaleUp,
-    TherapeuticInertia,
-)
+from vivarium_public_health.treatment import AbsoluteShift, LinearScaleUp, TherapeuticInertia
 
 __all__ = [
     __author__,
