--- src/vivarium_public_health/exposure/__init__.py
@@ -1,2 +1,0 @@
-from .effect import ExposureEffect
-from .exposure import Exposure
