+++ src/vivarium_public_health/results/__init__.py
@@ -1,7 +1,6 @@
 from .columns import COLUMNS
 from .disability import DisabilityObserver
 from .disease import DiseaseObserver
-from .intervention import CategoricalInterventionObserver
 from .mortality import MortalityObserver
 from .observer import PublicHealthObserver
 from .risk import CategoricalRiskObserver
