+++ tests/risks/test_data_transformations.py
@@ -1,8 +1,8 @@
 import pandas as pd
 import pytest
 
-from vivarium_public_health.exposure.distributions import DichotomousDistribution
 from vivarium_public_health.risks.data_transformations import _rebin_relative_risk_data
+from vivarium_public_health.risks.distributions import DichotomousDistribution
 
 
 @pytest.mark.parametrize(
@@ -24,13 +24,11 @@ def test__rebin_exposure_data(rebin_categories, rebinned_values):
             "value": [0.5] * 4 + [0.2] * 4 + [0.3] * 4,
         }
     )
-    rebinned_df = DichotomousDistribution._rebin_exposure_data(
-        df, rebin_categories, "exposed"
-    )
+    rebinned_df = DichotomousDistribution._rebin_exposure_data(df, rebin_categories)
 
     assert rebinned_df.shape == (4, 4)
     assert (rebinned_df.value == rebinned_values[0]).all()
-    assert (rebinned_df["parameter"] == "exposed").all()
+    assert (rebinned_df["parameter"] == "cat1").all()
 
 
 @pytest.mark.parametrize(
