+++ tests/risks/test_effect.py
@@ -4,13 +4,14 @@ import numpy as np
 import pandas as pd
 import pytest
 from layered_config_tree import LayeredConfigTree
-from vivarium import InteractiveContext
+from vivarium import Component, InteractiveContext
 from vivarium.framework.engine import Builder
 from vivarium.framework.event import Event
 from vivarium.framework.population import SimulantData
 from vivarium.testing_utilities import TestPopulation
 
 from vivarium_public_health.disease import SI
+from vivarium_public_health.risks import RiskEffect
 from vivarium_public_health.risks.base_risk import Risk
 
 #
@@ -511,7 +512,7 @@ def test_rr_sources(rr_source, rr_value, dichotomous_risk, base_config, base_plu
 custom_exposure_values = [0.5, 1, 1.5, 1.75, 2, 3, 4, 5, 5.5, 10]
 
 
-class CustomExposureRisk(Risk):
+class CustomExposureRisk(Component):
     """Risk where we define the exposure manually."""
 
     @property
@@ -523,7 +524,7 @@ class CustomExposureRisk(Risk):
         return [self.exposure_column_name]
 
     def __init__(self, risk: str):
-        super().__init__(risk)
+        super().__init__()
         self.risk = EntityString(risk)
         self.exposure_column_name = f"{self.risk.name}_exposure"
 
@@ -540,9 +541,6 @@ class CustomExposureRisk(Risk):
         builder.value.register_value_producer(
             f"{self.risk.name}.exposure",
             source=self.get_exposure,
-        )
-        self.randomness = builder.randomness.get_stream(
-            self.randomness_stream_name, component=self
         )
 
     def get_exposure(self, index: pd.Index) -> pd.Series:
@@ -562,6 +560,7 @@ class CustomExposureRisk(Risk):
 def test_non_loglinear_effect(rr_parameter_data, error_message, base_config, base_plugins):
     risk = CustomExposureRisk("risk_factor.test_risk")
     effect = NonLogLinearRiskEffect(risk.name, "cause.test_cause.incidence_rate")
+
     risk_effect_rrs = [2.0, 2.4, 4.0]
     rr_data = pd.DataFrame(
         {
@@ -635,7 +634,7 @@ def test_relative_risk_pipeline(dichotomous_risk, base_config, base_plugins):
     sim = _setup_risk_effect_simulation(base_config, base_plugins, risk, effect, data)
     pop = sim.get_population()
 
-    expected_pipeline_name = f"{effect.entity.name}_on_{effect.target.name}.relative_risk"
+    expected_pipeline_name = f"{effect.risk.name}_on_{effect.target.name}.relative_risk"
     assert expected_pipeline_name in sim.list_values()
 
     rr_mapper = {
@@ -643,7 +642,7 @@ def test_relative_risk_pipeline(dichotomous_risk, base_config, base_plugins):
         "cat2": 1.0,
     }
     for exposure in rr_mapper:
-        exposure_pipeline = sim.get_value(f"{effect.entity.name}.exposure")(pop.index)
+        exposure_pipeline = sim.get_value(f"{effect.risk.name}.exposure")(pop.index)
         exposure_idx = exposure_pipeline.loc[exposure_pipeline == exposure].index
         relative_risk = sim.get_value(expected_pipeline_name)(exposure_idx)
         assert (relative_risk == rr_mapper[exposure]).all()
